Let digitalBeamformedEnvelopeTrigger use the function phased_trigger as well. Implementing there the hilbert envelop
